+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ### 0.1.1
 
+* added default on JSONConfiguration
+* added DeclareProperties, so as to choose which action should be performed during setup (default to true)
+* removed ConsumerConfiguration, as it was considered duplication
 * abstracted consumer
 
 ### 0.1.0